+++ class/Purchase.rb
@@ -1,0 +1,7 @@
+class Purchase
+
+  def initialize(sales=0)
+    @sales = sales
+  end
+
+end
